Refactor ScatterPlot component, make it respond to resize
