+++ README.md
@@ -1,5 +1,7 @@
 A template project that uses Webpack and D3. Designed as a starting point for interactive data visualization projects that require JavaScript code to be organized across many files (as ES6 modules).
 
+The starter code here is from [Stylized Scatter Plot with Color Legend](https://bl.ocks.org/curran/ecb09f2605c7fbbadf0eeb75da5f0a6b).
+
 ## Development
 
 This project uses NPM and Webpack. To get started, clone the repository and install dependencies like this:
