+++ css/styles.css
@@ -1,19 +1,28 @@
-      body {
-        margin: 0px;
-      }
-      .domain {
-        display: none;
-      }
-      .tick line {
-        stroke: #C0C0BB;
-      }
-      .tick text, .legendCells text {
-        fill: #8E8883;
-        font-size: 28pt;
-        font-family: sans-serif;
-      }
-      .axis-label, .legend-label {
-        fill: #635F5D;
-        font-size: 50pt;
-        font-family: sans-serif;
-      }
+body {
+  margin: 0px;
+}
+.domain {
+  display: none;
+}
+.tick line {
+  stroke: #C0C0BB;
+}
+.tick text, .legendCells text {
+  fill: #8E8883;
+  font-size: 28pt;
+  font-family: sans-serif;
+}
+.axis-label, .legend-label {
+  fill: #635F5D;
+  font-size: 50pt;
+  font-family: sans-serif;
+}
+
+/* Make the chart container fill the page using CSS. */
+#visualization {
+  position: fixed;
+  left: 0px;
+  right: 0px;
+  top: 0px;
+  bottom: 0px;
+}
