+++ index.html
@@ -9,7 +9,9 @@
     <link rel="stylesheet" href="css/styles.css">
   </head>
   <body>
-    <svg width="960" height="500"></svg>
+    <div id="visualization">
+      <svg></svg>
+    </div>
     <script src="./dist/bundle.js"></script>
   </body>
 </html>
