+++ src/index.js
@@ -1,3 +1,5 @@
+import scatterPlot from './scatterPlot'
+
 const xValue = d => d.sepalLength;
 const xLabel = 'Sepal Length';
 const yValue = d => d.petalLength;
@@ -6,59 +8,9 @@ const colorValue = d => d.species;
 const colorLabel = 'Species';
 const margin = { left: 120, right: 300, top: 20, bottom: 120 };
 
-const svg = d3.select('svg');
-const width = svg.attr('width');
-const height = svg.attr('height');
-const innerWidth = width - margin.left - margin.right;
-const innerHeight = height - margin.top - margin.bottom;
-
-const g = svg.append('g')
-    .attr('transform', `translate(${margin.left},${margin.top})`);
-const xAxisG = g.append('g')
-    .attr('transform', `translate(0, ${innerHeight})`);
-const yAxisG = g.append('g');
-const colorLegendG = g.append('g')
-    .attr('transform', `translate(${innerWidth + 60}, 150)`);
-
-xAxisG.append('text')
-    .attr('class', 'axis-label')
-    .attr('x', innerWidth / 2)
-    .attr('y', 100)
-    .text(xLabel);
-
-yAxisG.append('text')
-    .attr('class', 'axis-label')
-    .attr('x', -innerHeight / 2)
-    .attr('y', -60)
-    .attr('transform', `rotate(-90)`)
-    .style('text-anchor', 'middle')
-    .text(yLabel);
-
-colorLegendG.append('text')
-    .attr('class', 'legend-label')
-    .attr('x', -30)
-    .attr('y', -40)
-    .text(colorLabel);
-
-const xScale = d3.scaleLinear();
-const yScale = d3.scaleLinear();
-const colorScale = d3.scaleOrdinal()
-  .range(d3.schemeCategory10);
-
-const xAxis = d3.axisBottom()
-  .scale(xScale)
-  .tickPadding(15)
-  .tickSize(-innerHeight);
-
-const yAxis = d3.axisLeft()
-  .scale(yScale)
-  .ticks(5)
-  .tickPadding(15)
-  .tickSize(-innerWidth);
-
-const colorLegend = d3.legendColor()
-  .scale(colorScale)
-  .shape('circle');
+const visualization = d3.select('#visualization');
+const visualizationDiv = visualization.node();
+const svg = visualization.select('svg');
 
 const row = d => {
   d.petalLength = +d.petalLength;
@@ -69,27 +21,30 @@ const row = d => {
 };
 
 d3.csv('data/iris.csv', row, data => {
-  xScale
-    .domain(d3.extent(data, xValue))
-    .range([0, innerWidth])
-    .nice();
 
-  yScale
-    .domain(d3.extent(data, yValue))
-    .range([innerHeight, 0])
-    .nice();
+  const render = () => {
 
-  g.selectAll('circle').data(data)
-    .enter().append('circle')
-      .attr('cx', d => xScale(xValue(d)))
-      .attr('cy', d => yScale(yValue(d)))
-      .attr('fill', d => colorScale(colorValue(d)))
-      .attr('fill-opacity', 0.6)
-      .attr('r', 8);
+    // Extract the width and height that was computed by CSS.
+    svg
+      .attr('width', visualizationDiv.clientWidth)
+      .attr('height', visualizationDiv.clientHeight);
 
-  xAxisG.call(xAxis);
-  yAxisG.call(yAxis);
-  colorLegendG.call(colorLegend)
-    .selectAll('.cell text')
-      .attr('dy', '0.1em');
+    // Render the scatter plot.
+    scatterPlot(svg, {
+      data,
+      xValue,
+      xLabel,
+      yValue,
+      yLabel,
+      colorValue,
+      colorLabel,
+      margin
+    });
+  }
+
+  // Draw for the first time to initialize.
+  render();
+
+  // Redraw based on the new size whenever the browser window is resized.
+  window.addEventListener('resize', render);
 });
