+++ src/scatterPlot.js
@@ -1,0 +1,110 @@
+const xScale = d3.scaleLinear();
+const yScale = d3.scaleLinear();
+const colorScale = d3.scaleOrdinal()
+  .range(d3.schemeCategory10);
+
+const xAxis = d3.axisBottom()
+  .scale(xScale)
+  .tickPadding(15);
+
+const yAxis = d3.axisLeft()
+  .scale(yScale)
+  .ticks(5)
+  .tickPadding(15);
+
+const colorLegend = d3.legendColor()
+  .scale(colorScale)
+  .shape('circle');
+
+export default function (svg, props) {
+  const { 
+    data,
+    xValue,
+    xLabel,
+    yValue,
+    yLabel,
+    colorValue,
+    colorLabel,
+    margin
+  } = props;
+
+  const width = svg.attr('width');
+  const height = svg.attr('height');
+  const innerWidth = width - margin.left - margin.right;
+  const innerHeight = height - margin.top - margin.bottom;
+
+  xAxis.tickSize(-innerHeight);
+  yAxis.tickSize(-innerWidth);
+
+  let g = svg.selectAll('.container').data([null]);
+  const gEnter = g.enter().append('g').attr('class', 'container');
+  g = gEnter
+    .merge(g)
+      .attr('transform', `translate(${margin.left},${margin.top})`);
+
+  const xAxisGEnter = gEnter.append('g').attr('class', 'x-axis');
+  const xAxisG = xAxisGEnter
+    .merge(g.select('.x-axis'))
+      .attr('transform', `translate(0, ${innerHeight})`);
+
+  const yAxisGEnter = gEnter.append('g').attr('class', 'y-axis');
+  const yAxisG = yAxisGEnter.merge(g.select('.y-axis'));
+
+  const colorLegendGEnter = gEnter.append('g').attr('class', 'legend');
+  const colorLegendG = colorLegendGEnter
+    .merge(g.select('.legend'))
+      .attr('transform', `translate(${innerWidth + 60}, 150)`);
+
+  xAxisGEnter
+    .append('text')
+      .attr('class', 'axis-label')
+      .attr('y', 100)
+    .merge(xAxisG.select('.axis-label'))
+      .attr('x', innerWidth / 2)
+      .text(xLabel);
+
+  yAxisGEnter
+    .append('text')
+      .attr('class', 'axis-label')
+      .attr('y', -60)
+      .style('text-anchor', 'middle')
+    .merge(yAxisG.select('.axis-label'))
+      .attr('x', -innerHeight / 2)
+      .attr('transform', `rotate(-90)`)
+      .text(yLabel);
+
+  colorLegendGEnter
+    .append('text')
+      .attr('class', 'legend-label')
+      .attr('x', -30)
+      .attr('y', -40)
+    .merge(colorLegendG.select('legend-label'))
+      .text(colorLabel);
+
+  xScale
+    .domain(d3.extent(data, xValue))
+    .range([0, innerWidth])
+    .nice();
+
+  yScale
+    .domain(d3.extent(data, yValue))
+    .range([innerHeight, 0])
+    .nice();
+
+  const circles = g.selectAll('.mark').data(data);
+  circles
+    .enter().append('circle')
+      .attr('class', 'mark')
+      .attr('fill-opacity', 0.6)
+      .attr('r', 8)
+    .merge(circles)
+      .attr('cx', d => xScale(xValue(d)))
+      .attr('cy', d => yScale(yValue(d)))
+      .attr('fill', d => colorScale(colorValue(d)));
+
+  xAxisG.call(xAxis);
+  yAxisG.call(yAxis);
+  colorLegendG.call(colorLegend)
+    .selectAll('.cell text')
+      .attr('dy', '0.1em');
+}
